refactor ProxyNodeFinder out of StreamrClient
