+++ packages/sdk/src/StreamrClient.ts
@@ -9,7 +9,7 @@ import './utils/PatchTsyringe'
 import { DhtAddress } from '@streamr/dht'
 import { ProxyDirection, StreamPartDeliveryOptions } from '@streamr/trackerless-network'
 import { DEFAULT_PARTITION_COUNT, EthereumAddress, HexString, Logger, StreamID, 
-    TheGraphClient, toEthereumAddress, toStreamPartID, toUserId } from '@streamr/utils'
+    TheGraphClient, toEthereumAddress, toUserId } from '@streamr/utils'
 import type { Overrides } from 'ethers'
 import EventEmitter from 'eventemitter3'
 import merge from 'lodash/merge'
@@ -31,7 +31,7 @@ import { MetricsPublisher } from './MetricsPublisher'
 import { NetworkNodeFacade } from './NetworkNodeFacade'
 import { RpcProviderSource } from './RpcProviderSource'
 import { Stream } from './Stream'
-import { DEFAULT_PARTITION, StreamIDBuilder } from './StreamIDBuilder'
+import { StreamIDBuilder } from './StreamIDBuilder'
 import { StreamMetadata, getPartitionCount } from './StreamMetadata'
 import { ChainEventPoller } from './contracts/ChainEventPoller'
 import { ContractFactory } from './contracts/ContractFactory'
@@ -62,8 +62,7 @@ import { convertPeerDescriptorToNetworkPeerDescriptor, createTheGraphClient } fr
 import { createIdentityFromConfig } from './identity/IdentityMapping'
 import { assertCompliantIdentity } from './utils/encryptionCompliance'
 import { SponsorshipFactory } from './contracts/SponsorshipFactory'
-import sample from 'lodash/sample'
-import shuffle from 'lodash/shuffle'
+import { ProxyNodeFinder } from './ProxyNodeFinder'
 
 // TODO: this type only exists to enable tsdoc to generate proper documentation
 export type SubscribeOptions = StreamDefinition & ExtraSubscribeOptions
@@ -115,6 +114,7 @@ export class StreamrClient {
     private readonly eventEmitter: StreamrClientEventEmitter
     private readonly destroySignal: DestroySignal
     private readonly loggerFactory: LoggerFactory
+    private readonly proxyNodeFinder: ProxyNodeFinder
 
     constructor(
         config: StreamrClientConfig = {},
@@ -150,6 +150,7 @@ export class StreamrClient {
         this.eventEmitter = container.resolve<StreamrClientEventEmitter>(StreamrClientEventEmitter)
         this.destroySignal = container.resolve<DestroySignal>(DestroySignal)
         this.loggerFactory = container.resolve<LoggerFactory>(LoggerFactory)
+        this.proxyNodeFinder = container.resolve<ProxyNodeFinder>(ProxyNodeFinder)
         container.resolve<PublisherKeyExchange>(PublisherKeyExchange) // side effect: activates publisher key exchange
         container.resolve<MetricsPublisher>(MetricsPublisher) // side effect: activates metrics publisher
         container.resolve<SponsorshipFactory>(SponsorshipFactory) // side effect: activates sponsorship event listeners
@@ -807,81 +808,7 @@ export class StreamrClient {
         maxQueryResults: number = 100,
         maxHeartbeatAgeHours: number = 24
     ): Promise<NetworkPeerDescriptor[]> {
-        const [streamId, partition] = await this.streamIdBuilder.toStreamPartElements(streamDefinition)
-        logger.debug(`findProxyNodes: trying to find ${numberOfProxies} proxy nodes for stream ${streamId} partition ${partition}`)
-
-        // Find active operators on the stream
-        const foundOperators = await this.operatorRegistry.findOperatorsOnStream(streamId, maxQueryResults, maxHeartbeatAgeHours)
-        logger.debug(`findProxyNodes: found ${foundOperators.length} operators on stream ${streamId}`)
-        
-        // Throw early if we don't even have a chance of finding enough proxies (we accept one per operator)
-        if (foundOperators.length < numberOfProxies) {
-            throw new Error(`Not enough operators found for stream ${streamId}: found ${
-                foundOperators.length} operators, but ${numberOfProxies} are required`)
-        }
-        
-        const shuffledOperators = shuffle(foundOperators)
-        const foundProxies: NetworkPeerDescriptor[] = []
-
-        // This shared promise is used to signal that enough proxies have been found
-        // It's a speed optimization to avoid waiting for all workers to complete if we already have enough nodes
-        let signalEnoughProxiesFound: () => void = () => {}
-        const enoughProxiesFoundPromise = new Promise<void>((resolve) => {
-            signalEnoughProxiesFound = resolve
-        })
-
-        // Try to find numOfProxies nodes from the shuffled operators
-        // We search in parallel, with parallelism = numOfProxies
-        // If there's an error, try the next operator
-        let operatorIndex = 0
-        const startNodeDiscoveryWorker = async (workerIndex: number) => {
-            while (operatorIndex < shuffledOperators.length && foundProxies.length < numberOfProxies) {
-                const operator = shuffledOperators[operatorIndex]
-                operatorIndex++
-                const streamPartId = toStreamPartID(streamId, partition ?? DEFAULT_PARTITION)
-                try {
-                    const nodes = await this.node.discoverOperators(operator.peerDescriptor, streamPartId)
-
-                    if (nodes.length > 0) {
-                        // One more check to make sure the other racing threads didn't already find enough nodes
-                        if (foundProxies.length < numberOfProxies) {
-                            logger.debug(`findProxyNodes (worker ${workerIndex}): found ${nodes.length} nodes for operator ${
-                                operator.operatorId} and streamPartId ${streamPartId}`)
-        
-                            foundProxies.push(sample(nodes)!)
-
-                            if (foundProxies.length >= numberOfProxies) {
-                                // Signal that enough proxies have been found, this will resolve the race
-                                // and ignore any further results from the other workers
-                                signalEnoughProxiesFound()
-                            }
-                        } else {
-                            // silently drop the result, enough have been found already
-                        }
-                    } else {
-                        logger.debug(`findProxyNodes (worker ${workerIndex}): no nodes found for operator ${
-                            operator.operatorId} and streamPartId ${streamPartId}`)
-                    }
-                } catch (error) {
-                    logger.error(`findProxyNodes (worker ${workerIndex}): error discovering nodes for operator ${
-                        operator.operatorId}: ${(error as Error).message}`)
-                }
-            }
-        }
-
-        // Start numberOfProxies discovery workers in parallel
-        // The race will resolve as soon as they signal that they 
-        // found enough nodes (success) or the workers complete (failure)
-        await Promise.race([
-            Promise.all(Array.from({ length: numberOfProxies }, (_, index) => startNodeDiscoveryWorker(index))),
-            enoughProxiesFoundPromise
-        ])
-
-        if (foundProxies.length < numberOfProxies) {
-            throw new Error(`Not enough proxy nodes were resolved for stream ${streamId}: found ${
-                foundProxies.length} nodes, but ${numberOfProxies} are required`)
-        }
-        return foundProxies
+        return this.proxyNodeFinder.find(streamDefinition, numberOfProxies, maxQueryResults, maxHeartbeatAgeHours)
     }
 
     /**
